copilot/cpu/aws-copilot-inference-service/requirements.txt
aiohttp==3.7.4
onnxruntime==1.8.1
Pillow==9.4.0